requirements.txt
streamlit
opencv-python
ultralytics
python-dotenv
requests
Pillow
numpy
imgurpython
google-generativeai